testing

+++ src/activation.rs
@@ -1,4 +1,4 @@
-use std::f32::consts::E;
+use std::f64::consts::E;
 
 #[derive(Clone, Copy)]
 pub enum Act {
@@ -7,13 +7,13 @@ pub enum Act {
     Lin
 }
 
-fn sigmoid(n: f32) -> f32 {
+fn sigmoid(n: f64) -> f64 {
     1.0 / (1.0 + E.powf(-n))
 }
 
 impl Act {
     /// Applies non-linearity function to `n`
-    pub fn value(&self, n: f32) -> f32 {
+    pub fn value(&self, n: f64) -> f64 {
         match self {
             Act::Tanh => n.tanh(),
             Act::Sig  => sigmoid(n),
@@ -22,7 +22,7 @@ impl Act {
     }
 
     /// Applies non-linearity derivative to `n`
-    pub fn deriv(&self, n: f32) -> f32 {
+    pub fn deriv(&self, n: f64) -> f64 {
         match self {
             Act::Tanh => 1.0  - n.tanh().powi(2),
             Act::Sig  => {
