+++ src/cost.rs
@@ -6,14 +6,14 @@ pub enum Cost {
 
 impl Cost {
     /// Applies cost function to `err`
-    pub fn value(&self, err: f32) -> f32 {
+    pub fn value(&self, err: f64) -> f64 {
         match self {
             Cost::MSE => err.powi(2)
         }
     }
 
     /// Applies cost derivative to `err`
-    pub fn deriv(&self, err: f32) -> f32 {
+    pub fn deriv(&self, err: f64) -> f64 {
         match self {
             Cost::MSE => 2.0 * err
         }
